LayoutTests/http/tests/appcache/simple-expected.txt

ALERT: whee cached!
This tests that the application cache works by loading a simple text file from the cache using XMLHttpRequest.
